updated API for events
